Rename staff_tenant file to match class name
